KR - Optimisation du code

+++ src/Controller/AdminEmailsController.php
@@ -4,7 +4,9 @@ namespace App\Controller;
 
 use App\Entity\EmailsList;
 use App\Form\EmailAdminFormType;
+use App\Repository\EmailsListRepository;
 use Cocur\Slugify\Slugify;
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,10 +16,8 @@ use Symfony\Component\Routing\Annotation\Route;
 class AdminEmailsController extends AbstractController
 {
     #[Route('/admin/emails', name: 'app_admin_emails')]
-    public function index(ManagerRegistry $doctrine): Response
+    public function index(EmailsListRepository $emailsRepo): Response
     {
-        $em = $doctrine->getManager();
-        $emailsRepo = $em->getRepository(EmailsList::class);
         $emails = $emailsRepo->findAll();
 
         return $this->render('admin_emails/index.html.twig', [
@@ -26,14 +26,13 @@ class AdminEmailsController extends AbstractController
     }
 
     #[Route('/admin/emails/ajouter', name: 'app_admin_emails_create')]
-    public function create(Request $request, ManagerRegistry $doctrine): Response
+    public function create(Request $request, EntityManagerInterface $em): Response
     {
         $form = $this->createForm(EmailAdminFormType::class);
         $form->handleRequest($request);
         
         if ($form->isSubmitted() && $form->isValid()) {
             // Récupération des valeurs
-            $email = new EmailsList();
             $emailName = $form->get('email_name')->getData();
             $emailContent = $form->get('email_content')->getData();
 
@@ -42,16 +41,15 @@ class AdminEmailsController extends AbstractController
             $emailId = $slugify->slugify($emailName);
 
             // Création du template
-            $file = fopen("../templates/emails/" . $emailId . '.html.twig', 'w');
-            fwrite($file, $emailContent);
-            fclose($file);
+            $filePath = "../templates/emails/" . $emailId . '.html.twig';
+            file_put_contents($filePath, $emailContent);
 
             // Envoi des données vers la BDD
+            $email = new EmailsList();
             $email->setEmailName($emailName);
             $email->setEmailId($emailId);
-            $entityManager = $doctrine->getManager();
-            $entityManager->persist($email);
-            $entityManager->flush();
+            $em->persist($email);
+            $em->flush();
         }
 
         return $this->render('admin_emails/email_manager.html.twig', [
@@ -62,14 +60,21 @@ class AdminEmailsController extends AbstractController
     }
 
     #[Route('/admin/email/{emailId}', name: 'app_admin_emails_update')]
-    public function update(Request $request, ManagerRegistry $doctrine, string $emailId)
+    public function update(Request $request, EntityManagerInterface $em, ManagerRegistry $doctrine, string $emailId)
     {
         $form = $this->createForm(EmailAdminFormType::class);
         $form->handleRequest($request);
 
         // Récupération de l'E-mail
-        $entityManager = $doctrine->getManager();
-        $email = $entityManager->getRepository(EmailsList::class)->findOneBy(['email_id' => $emailId]);
+        $emailRepository = $em->getRepository(EmailsList::class);
+        $email = $emailRepository->findOneBy(['email_id' => $emailId]);
+
+        if (!$email) {
+            throw $this->createNotFoundException('Email not found.');
+        }
+
+        $form = $this->createForm(EmailAdminFormType::class, $email);
+        $form->handleRequest($request);
         $emailContent = file_get_contents("../templates/emails/" . $emailId . ".html.twig");
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -77,16 +82,15 @@ class AdminEmailsController extends AbstractController
             $emailContent = $form->get('email_content')->getData();
             
             // Modification du fichier
-            unlink("../templates/emails/" . $email->getEmailId() . '.html.twig');
-            $file = fopen("../templates/emails/" . $email->getEmailId() . '.html.twig', 'w');
-            fwrite($file, $emailContent);
-            fclose($file);
+            $filePath = "../templates/emails/" . $email->getEmailId() . '.html.twig';
+            if (file_exists($filePath)) {
+                unlink($filePath);
+            }
+            file_put_contents($filePath, $emailContent);
 
             // Envoi des données vers la BDD
             $email->setEmailName($emailName);
-            $entityManager = $doctrine->getManager();
-            $entityManager->persist($email);
-            $entityManager->flush();
+            $em->flush();
         }
 
         return $this->render('admin_emails/email_manager.html.twig', [
@@ -97,16 +101,23 @@ class AdminEmailsController extends AbstractController
     }
 
     #[Route('/admin/email/{emailId}/delete', name: 'app_admin_emails_delete')]
-    public function delete(Request $request, ManagerRegistry $doctrine, string $emailId)
+    public function delete(Request $request, EntityManagerInterface $entityManager, string $emailId)
     {
         // Récupération de l'E-mail
-        $entityManager = $doctrine->getManager();
-        $email = $entityManager->getRepository(EmailsList::class)->findOneBy(['email_id' => $emailId]);
+        $emailRepository = $entityManager->getRepository(EmailsList::class);
+        $email = $emailRepository->findOneBy(['email_id' => $emailId]);
+
+        if (!$email) {
+            throw $this->createNotFoundException("L'E-Mail n'a pas été trouvé");
+        }
         
         // Suppression du fichier
-        unlink("../templates/emails/" . $email->getEmailId() . ".html.twig");
+        $filePath = "../templates/emails/" . $email->getEmailId() . ".html.twig";
+        if (file_exists($filePath)) {
+            unlink($filePath);
+        }
 
-        // Suppresison dans la DB
+        // Suppression dans la DB
         $entityManager->remove($email);
         $entityManager->flush();
